refactor(core): integrate penelope into api_swedeb
Closes #300
