+++ api_swedeb/api/services/corpus_loader.py
@@ -17,11 +17,11 @@ import pandas as pd
 
 from api_swedeb.core import person_codecs as md
 from api_swedeb.core.configuration import ConfigValue
+from api_swedeb.core.dtm import IVectorizedCorpus
 from api_swedeb.core.load import load_dtm_corpus, load_speech_index
 from api_swedeb.core.speech_repository import SpeechRepository
 from api_swedeb.core.speech_store import SpeechStore
 from api_swedeb.core.utility import Lazy
-from penelope.corpus import IVectorizedCorpus
 
 
 class CorpusLoader:
