+++ api_swedeb/core/speech_index.py
@@ -2,9 +2,9 @@ from typing import Any
 
 import pandas as pd
 
+from api_swedeb.core.common.utility import PropertyValueMaskingOpts
+from api_swedeb.core.dtm import VectorizedCorpus
 from api_swedeb.core.utility import filter_by_opts
-from penelope.corpus import VectorizedCorpus
-from penelope.utility import PropertyValueMaskingOpts
 
 
 def word_in_vocabulary(corpus: VectorizedCorpus, word: str) -> str | None:
