--- penelope/__init__.py
@@ -1,1 +1,0 @@
-__version__ = "0.8.3"
