--- penelope/corpus/__init__.py
@@ -1,8 +1,0 @@
-# type: ignore
-from .dtm import (
-    IVectorizedCorpus,
-    VectorizedCorpus,
-    find_matching_words_in_vocabulary,
-    load_metadata,
-    store_metadata,
-)
